app/exec/Main.hs, src/Kore/Exec.hs: Refactor kore-exec into a library
